[WIP] Refactorings (#452)
Split code in different modules
