+++ .rubocop.yml
@@ -10,30 +10,32 @@ Style/FrozenStringLiteralComment:
 Style/EmptyMethod:
   Enabled: false
 
-Style/VariableNumber:
-  Enabled: false
-
 Style/ClassAndModuleChildren:
   Enabled: false
 
 Metrics/LineLength:
   Max: 120
 
-Metrics/ClassLength:
-  Exclude:
-    - 'lib/cancan/controller_resource.rb'
-    - 'lib/cancan/rule.rb'
-
-Metrics/ModuleLength:
-  Exclude:
-    - "**/*_spec.rb"
-    - 'lib/cancan/ability.rb'
-    - 'lib/cancan/model_adapters/active_record_adapter.rb'
-
 Metrics/BlockLength:
   Exclude:
     - 'lib/cancan/matchers.rb'
     - '**/*_spec.rb'
+
+# TODO
+# Offense count: 2
+# Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist.
+# NamePrefix: is_, has_, have_
+# NamePrefixBlacklist: is_, has_, have_
+# NameWhitelist: is_a?
+Style/PredicateName:
+  Exclude:
+    - 'spec/**/*'
+    - 'lib/cancan/ability.rb'
+
+# TODO
+Lint/AmbiguousBlockAssociation:
+  Enabled: false
+
 
 AllCops:
   TargetRubyVersion: 2.0
