--- .rubocop_todo.yml
@@ -1,91 +1,0 @@
-# This configuration was generated by
-# `rubocop --auto-gen-config`
-# on 2017-03-26 15:25:15 +0200 using RuboCop version 0.46.0.
-# The point is for the user to remove these configuration records
-# one by one as the offenses are removed from the code base.
-# Note that changes in the inspected code, or installation of new
-# versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 12
-Metrics/AbcSize:
-  Max: 21
-
-# Offense count: 3
-# Configuration parameters: CountComments.
-Metrics/BlockLength:
-  Max: 58
-
-# Offense count: 4
-Metrics/CyclomaticComplexity:
-  Max: 9
-
-# Offense count: 13
-# Configuration parameters: CountComments.
-Metrics/MethodLength:
-  Max: 21
-
-# Offense count: 4
-Metrics/PerceivedComplexity:
-  Max: 10
-
-# Offense count: 3
-# Cop supports --auto-correct.
-# Configuration parameters: EnforcedStyle, SupportedStyles, ProceduralMethods, FunctionalMethods, IgnoredMethods.
-# SupportedStyles: line_count_based, semantic, braces_for_chaining
-# ProceduralMethods: benchmark, bm, bmbm, create, each_with_object, measure, new, realtime, tap, with_object
-# FunctionalMethods: let, let!, subject, watch
-# IgnoredMethods: lambda, proc, it
-Style/BlockDelimiters:
-  Exclude:
-    - 'spec/cancan/matchers_spec.rb'
-
-# Offense count: 4
-# Configuration parameters: ExpectMatchingDefinition, Regex, IgnoreExecutableScripts.
-Style/FileName:
-  Exclude:
-    - 'spec/cancan/inherited_resource_spec_BACKUP_24010.rb'
-    - 'spec/cancan/inherited_resource_spec_BASE_24010.rb'
-    - 'spec/cancan/inherited_resource_spec_LOCAL_24010.rb'
-    - 'spec/cancan/inherited_resource_spec_REMOTE_24010.rb'
-
-# Offense count: 2
-# Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist.
-# NamePrefix: is_, has_, have_
-# NamePrefixBlacklist: is_, has_, have_
-# NameWhitelist: is_a?
-Style/PredicateName:
-  Exclude:
-    - 'spec/**/*'
-    - 'lib/cancan/ability.rb'
-
-# Offense count: 1
-# Cop supports --auto-correct.
-# Configuration parameters: EnforcedStyle, SupportedStyles.
-# SupportedStyles: space, no_space
-Style/SpaceAroundEqualsInParameterDefault:
-  Exclude:
-    - 'spec/cancan/inherited_resource_spec_BASE_24010.rb'
-
-# Offense count: 4
-# Cop supports --auto-correct.
-# Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces, SupportedStyles.
-# SupportedStyles: space, no_space, compact
-Style/SpaceInsideHashLiteralBraces:
-  Exclude:
-    - 'spec/cancan/inherited_resource_spec_BASE_24010.rb'
-
-# Offense count: 1
-# Cop supports --auto-correct.
-Style/TrailingWhitespace:
-  Exclude:
-    - 'lib/cancan.rb'
-
-
-# has a bug
-Style/FileName:
-  Exclude:
-    - 'Appraisals'
-
-# disagree
-Lint/AmbiguousBlockAssociation:
-  Enabled: false
