+++ gemfiles/activerecord_4.2.gemfile
@@ -2,10 +2,10 @@
 
 source "https://rubygems.org"
 
-gem "activerecord", "~> 4.2.0", :require => "active_record"
-gem "activesupport", "~> 4.2.0", :require => "active_support/all"
-gem "actionpack", "~> 4.2.0", :require => "action_pack"
-gem "nokogiri", "~> 1.6.8", :require => "nokogiri"
+gem "activerecord", "~> 4.2.0", require: "active_record"
+gem "activesupport", "~> 4.2.0", require: "active_support/all"
+gem "actionpack", "~> 4.2.0", require: "action_pack"
+gem "nokogiri", "~> 1.6.8", require: "nokogiri"
 
 platforms :jruby do
   gem "activerecord-jdbcsqlite3-adapter"
@@ -17,4 +17,4 @@ platforms :ruby, :mswin, :mingw do
   gem "pg"
 end
 
-gemspec :path => "../"
+gemspec path: "../"
