+++ gemfiles/activerecord_5.1.0.gemfile
@@ -2,9 +2,9 @@
 
 source "https://rubygems.org"
 
-gem "activerecord", "~> 5.1.0", :require => "active_record"
-gem "activesupport", "~> 5.1.0", :require => "active_support/all"
-gem "actionpack", "~> 5.1.0", :require => "action_pack"
+gem "activerecord", "~> 5.1.0", require: "active_record"
+gem "activesupport", "~> 5.1.0", require: "active_support/all"
+gem "actionpack", "~> 5.1.0", require: "action_pack"
 
 platforms :jruby do
   gem "activerecord-jdbcsqlite3-adapter"
@@ -16,4 +16,4 @@ platforms :ruby, :mswin, :mingw do
   gem "pg"
 end
 
-gemspec :path => "../"
+gemspec path: "../"
