+++ lib/cancan.rb
@@ -10,6 +10,6 @@ require 'cancan/model_adapters/abstract_adapter'
 require 'cancan/model_adapters/default_adapter'
 
 if defined? ActiveRecord
-  require 'cancan/model_adapters/active_record_adapter'  
+  require 'cancan/model_adapters/active_record_adapter'
   require 'cancan/model_adapters/active_record_4_adapter'
 end
