+++ spec/cancan/matchers_spec.rb
@@ -11,23 +11,23 @@ describe 'be_able_to' do
 
     it 'reports a nice failure message for should' do
       is_expected.to receive(:can?).with(:read, 123) { false }
-      expect {
+      expect do
         is_expected.to be_able_to(:read, 123)
-      }.to raise_error('expected to be able to :read 123')
+      end.to raise_error('expected to be able to :read 123')
     end
 
     it 'reports a nice failure message for should not' do
       is_expected.to receive(:can?).with(:read, 123) { true }
-      expect {
+      expect do
         is_expected.to_not be_able_to(:read, 123)
-      }.to raise_error('expected not to be able to :read 123')
+      end.to raise_error('expected not to be able to :read 123')
     end
 
     it 'delegates additional arguments to can? and reports in failure message' do
       is_expected.to receive(:can?).with(:read, 123, 456) { false }
-      expect {
+      expect do
         is_expected.to be_able_to(:read, 123, 456)
-      }.to raise_error('expected to be able to :read 123 456')
+      end.to raise_error('expected to be able to :read 123 456')
     end
   end
